update for passing some indo
